+++ src/power_grid_model/core/data_handling.py
@@ -11,12 +11,13 @@ from typing import Mapping
 
 import numpy as np
 
+from power_grid_model._utils import process_data_filter
 from power_grid_model.core.dataset_definitions import ComponentType, DatasetType
 from power_grid_model.core.power_grid_dataset import CConstDataset, CMutableDataset
-from power_grid_model.core.power_grid_meta import initialize_array, power_grid_meta_data
+from power_grid_model.core.power_grid_meta import initialize_array
 from power_grid_model.data_types import Dataset
 from power_grid_model.enum import CalculationType
-from power_grid_model.typing import ComponentAttributeMapping, _ComponentAttributeMappingDict
+from power_grid_model.typing import ComponentAttributeMapping
 
 
 def get_output_type(*, calculation_type: CalculationType, symmetric: bool) -> DatasetType:
@@ -126,66 +127,3 @@ def create_output_data(
             shape = (count,)
         result_dict[name] = initialize_array(output_type, name, shape=shape, empty=True)
     return result_dict
-
-
-def process_data_filter(
-    dataset_type: DatasetType,
-    data_filter: ComponentAttributeMapping,
-    available_components: list[ComponentType],
-) -> _ComponentAttributeMappingDict:
-    """Checks valid type for a mapping of componenet to attributes.
-    The default for `None` is row_based format while Ellipsis (...) is columnar format.
-    Also checks for any invalid component names and attribute names
-
-    Args:
-        dataset_type (DatasetType): the type of output that the user will see (as per the calculation options)
-        data_filter (ComponentAttributeMapping):  component to attribute mapping or list provided by user
-        available_components (list[ComponentType]):  all components available in model instance
-
-    Returns:
-        _ComponentAttributeMappingDict: processed component to attribute mapping
-    """
-    # limit all component count to user specified component types in output and convert to a dict
-    if data_filter is None:
-        data_filter = {k: None for k in available_components}
-    elif data_filter is Ellipsis:
-        data_filter = {k: Ellipsis for k in available_components}
-    elif isinstance(data_filter, (list, set)):
-        data_filter = {k: None for k in data_filter}
-    elif not isinstance(data_filter, dict) or not all(
-        attrs is None or attrs is Ellipsis or isinstance(attrs, (set, list)) for attrs in data_filter.values()
-    ):
-        raise ValueError(f"Invalid filter provided: {data_filter}")
-
-    validate_data_filter(data_filter, dataset_type)
-
-    return data_filter
-
-
-def validate_data_filter(data_filter: _ComponentAttributeMappingDict, dataset_type: DatasetType) -> None:
-    """Raise error if some specified components or attributes are unknown
-
-    Args:
-        data_filter (_ComponentAttributeMappingDict): Component to attribtue mapping
-        dataset_type (DatasetType):  Type of dataset
-
-    Raises:
-        ValueError: when the type for output_comoponent_types is incorrect
-        KeyError: with "unknown component" for any unknown components
-        KeyError: with "unknown attributes" for any unknown attributes for a known component
-    """
-    dataset_meta = power_grid_meta_data[dataset_type]
-    unknown_components = [x for x in data_filter if x not in dataset_meta]
-    if unknown_components:
-        raise KeyError(f"You have specified some unknown component types: {unknown_components}")
-
-    unknown_attributes = {}
-    for comp_name, attrs in data_filter.items():
-        if attrs is None or attrs is Ellipsis:
-            continue
-        diff = set(attrs).difference(dataset_meta[comp_name].dtype.names)
-        if diff != set():
-            unknown_attributes[comp_name] = diff
-
-    if unknown_attributes:
-        raise KeyError(f"You have specified some unknown attributes: {unknown_attributes}")
