+++ tests/unit/test_data_handling.py
@@ -5,7 +5,8 @@
 import numpy as np
 import pytest
 
-from power_grid_model.core.data_handling import create_output_data, process_data_filter
+from power_grid_model._utils import process_data_filter
+from power_grid_model.core.data_handling import create_output_data
 from power_grid_model.core.dataset_definitions import ComponentType as CT, DatasetType as DT
 from power_grid_model.core.power_grid_meta import initialize_array
